refactor: SavedVideoListView 분리 및 view 내부 함수 이름 변경
